voice-chatbot-hitl/requirements.txt
langgraph
langchain
langchain-openai
streamlit
python-dotenv
deepgram-sdk
audio-recorder-streamlit
requests
sqlite3